Adds check for betweenness of collinear points
After calculating the intersection between the line segment formed by
the center of project and the point that's being determined whether to
occlude, we check whether the intersection point is between the question
point and the center of projection.

If this interection point is not between these other two points, we
don't bother with the point_in_polygon check because the intersection is
"behind" in our view of the shape anyway.